% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_dataset.R
\name{clean_dataset}
\alias{clean_dataset}
\title{clean_dataset}
\usage{
clean_dataset(datasets)
}
\arguments{
\item{datasets}{a tibble object returned from list_datasets()}
}
\value{
a tibble
}
\description{
this function will clean the messy dataset from list_datasets.
}
\examples{
\dontrun{
datasets <- list_datasets(country = "Saudi Arabia")
clean_dataset(datasets)
}

}
