% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_attachments.R
\name{download_attachments}
\alias{download_attachments}
\title{Download attachment}
\usage{
download_attachments(attachments)
}
\arguments{
\item{attachments}{the return from  get_attachments()}
}
\value{
a tibble
}
\description{
Download attachment
}
\examples{
\dontrun{
datasets <- list_datasets(country = "Saudi Arabia")
datasetID <- datasets$dataset_id[1]
attachments <- get_attachments(datasetID)
download_attachments(attachments)
}

}
