% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attachments.R
\name{get_attachments}
\alias{get_attachments}
\title{Get list of all available attachments}
\usage{
get_attachments(datasetID)
}
\arguments{
\item{datasetID}{a Dataset ID. Can be found in the first coulmn of list_datasets() under the name of dataset_id}
}
\value{
a list
}
\description{
Get list of all available attachments
}
\examples{
\dontrun{
datasets<-list_datasets(country = "Saudi Arabia")
datasetID<-datasets$dataset_id[1]
get_attachments(datasetID)
}

}
