% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Get Datasets Records}
\usage{
get_dataset(datasetID)
}
\arguments{
\item{datasetID}{a Dataset ID. Can be found in the first coulmn of list_datasets() under the name of dataset_id}
}
\value{
a tibble
}
\description{
Get Datasets Records
}
\examples{
\dontrun{
get_dataset(datasetID)
}

}
