% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{List all Datasets}
\usage{
list_datasets(
  q = NULL,
  country = NULL,
  theme = NULL,
  keyword = NULL,
  operation = "and"
)
}
\arguments{
\item{q}{text search over all the meta}

\item{country}{Country name to filter datasets on}

\item{theme}{theme name to filter datasets on [look list_themes()]}

\item{keyword}{keword name to filter datasets on [look list_keywords()]}

\item{operation}{query operation [and , or(the Default value)]}
}
\value{
a tibble
}
\description{
List all Datasets
}
\examples{
\dontrun{
list_datasets(country = "Saudi Arabia")
list_datasets(q = "custom.country = 'Saudi Arabia'")
list_datasets(theme = "water", keyword = "Water", operation = "or")
}

}
